[REF/#290] 일기 관련 탭 공통 컴포넌트화
