+++ core/designsystem/src/main/java/com/hilingual/core/designsystem/model/DiaryContent.kt
@@ -13,28 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hilingual.presentation.diaryfeedback.model
+package com.hilingual.core.designsystem.model
 
 import androidx.compose.runtime.Immutable
-import com.hilingual.data.diary.model.DiaryRecommendExpressionModel
 import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.toImmutableList
+import kotlinx.collections.immutable.persistentListOf
 
 @Immutable
-internal data class RecommendExpressionUiModel(
-    val phraseId: Long,
-    val phraseType: ImmutableList<String>,
-    val phrase: String,
-    val explanation: String,
-    val reason: String,
-    val isMarked: Boolean = false
-)
-
-internal fun DiaryRecommendExpressionModel.toState() = RecommendExpressionUiModel(
-    phraseId = this.phraseId,
-    phraseType = this.phraseType.toImmutableList(),
-    phrase = this.phrase,
-    explanation = this.explanation,
-    reason = this.reason,
-    isMarked = this.isMarked
+data class DiaryContent(
+    val originalText: String = "",
+    val aiText: String = "",
+    val diffRanges: ImmutableList<Pair<Int, Int>> = persistentListOf(),
+    val imageUrl: String? = null
 )
