+++ presentation/diaryfeedback/src/main/java/com/hilingual/presentation/diaryfeedback/DiaryFeedbackUiState.kt
@@ -16,17 +16,45 @@
 package com.hilingual.presentation.diaryfeedback
 
 import androidx.compose.runtime.Immutable
-import com.hilingual.presentation.diaryfeedback.model.DiaryContentUiModel
-import com.hilingual.presentation.diaryfeedback.model.FeedbackContentUiModel
-import com.hilingual.presentation.diaryfeedback.model.RecommendExpressionUiModel
+import com.hilingual.core.designsystem.model.DiaryContent
+import com.hilingual.core.designsystem.model.FeedbackContent
+import com.hilingual.core.designsystem.model.RecommendExpression
+import com.hilingual.data.diary.model.DiaryContentModel
+import com.hilingual.data.diary.model.DiaryFeedbackModel
+import com.hilingual.data.diary.model.DiaryRecommendExpressionModel
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
+import kotlinx.collections.immutable.toImmutableList
 
 @Immutable
 internal data class DiaryFeedbackUiState(
     val writtenDate: String = "",
     val isPublished: Boolean = false,
-    val diaryContent: DiaryContentUiModel = DiaryContentUiModel(),
-    val feedbackList: ImmutableList<FeedbackContentUiModel> = persistentListOf(),
-    val recommendExpressionList: ImmutableList<RecommendExpressionUiModel> = persistentListOf()
+    val diaryContent: DiaryContent = DiaryContent(),
+    val feedbackList: ImmutableList<FeedbackContent> = persistentListOf(),
+    val recommendExpressionList: ImmutableList<RecommendExpression> = persistentListOf()
+)
+
+internal fun DiaryContentModel.toState() = DiaryContent(
+    originalText = this.originalText,
+    aiText = this.rewriteText,
+    diffRanges = this.diffRanges.map {
+        it.diffRange.first to it.diffRange.second
+    }.toImmutableList(),
+    imageUrl = this.imageUrl
+)
+
+internal fun DiaryFeedbackModel.toState() = FeedbackContent(
+    originalText = this.originalText,
+    feedbackText = this.rewriteText,
+    explain = this.explain
+)
+
+internal fun DiaryRecommendExpressionModel.toState() = RecommendExpression(
+    phraseId = this.phraseId,
+    phraseType = this.phraseType.toImmutableList(),
+    phrase = this.phrase,
+    explanation = this.explanation,
+    reason = this.reason,
+    isMarked = this.isMarked
 )
