--- presentation/diaryfeedback/src/main/java/com/hilingual/presentation/diaryfeedback/model/DiaryContentUiModel.kt
@@ -1,39 +1,0 @@
-/*
- * Copyright 2025 The Hilingual Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.hilingual.presentation.diaryfeedback.model
-
-import androidx.compose.runtime.Immutable
-import com.hilingual.data.diary.model.DiaryContentModel
-import kotlinx.collections.immutable.ImmutableList
-import kotlinx.collections.immutable.persistentListOf
-import kotlinx.collections.immutable.toImmutableList
-
-@Immutable
-internal data class DiaryContentUiModel(
-    val originalText: String = "",
-    val aiText: String = "",
-    val diffRanges: ImmutableList<Pair<Int, Int>> = persistentListOf(),
-    val imageUrl: String? = null
-)
-
-internal fun DiaryContentModel.toState() = DiaryContentUiModel(
-    originalText = this.originalText,
-    aiText = this.rewriteText,
-    diffRanges = this.diffRanges.map {
-        it.diffRange.first to it.diffRange.second
-    }.toImmutableList(),
-    imageUrl = this.imageUrl
-)
